nicolo/text_chunking-main/requirements.txt
langchain
langchain-openai
langchain-community
langchain-text-splitters
llama-index
umap-learn
hdbscan
python-dotenv
seaborn
datamapplot
black